+++ src/type_/expr.rs
@@ -21,12 +21,6 @@ pub struct ExprTyper<'a, 'b, 'c> {
     pub ungeneralised_function_used: bool,
 }
 
-impl<'a, 'b, 'c> Typer for ExprTyper<'a, 'b, 'c> {
-    fn with_environment<T>(&mut self, f: impl FnOnce(&mut Environment<'_, '_>) -> T) -> T {
-        f(self.environment)
-    }
-}
-
 impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
     pub fn new(environment: &'a mut Environment<'b, 'c>) -> Self {
         let mut hydrator = Hydrator::new();
@@ -66,6 +60,14 @@ impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
             .instantiate(t, ctx_level, ids, &self.hydrator)
     }
 
+    fn new_unbound_var(&mut self, level: usize) -> Arc<Type> {
+        self.environment.new_unbound_var(level)
+    }
+
+    fn unify(&mut self, t1: Arc<Type>, t2: Arc<Type>) -> Result<(), UnifyError> {
+        self.environment.unify(t1, t2)
+    }
+
     /// Crawl the AST, annotating each node with the inferred type or
     /// returning an error.
     ///
@@ -81,7 +83,10 @@ impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
                 location, value, ..
             } => self.infer_int(value, location),
 
-            UntypedExpr::Seq { first, then, .. } => self.infer_seq(*first, *then),
+            UntypedExpr::Sequence {
+                expressions,
+                location,
+            } => self.infer_seq(location, expressions),
 
             UntypedExpr::Tuple {
                 location, elems, ..
@@ -114,11 +119,19 @@ impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
                 location,
                 pattern,
                 value,
-                then,
                 kind,
                 annotation,
                 ..
-            } => self.infer_let(pattern, *value, *then, kind, &annotation, location),
+            } => self.infer_assignment(pattern, *value, kind, &annotation, location),
+
+            UntypedExpr::Try {
+                location,
+                pattern,
+                value,
+                then,
+                annotation,
+                ..
+            } => self.infer_try(pattern, *value, *then, &annotation, location),
 
             UntypedExpr::Case {
                 location,
@@ -329,28 +342,44 @@ impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
         })
     }
 
-    fn infer_seq(&mut self, first: UntypedExpr, then: UntypedExpr) -> Result<TypedExpr, Error> {
-        if first.is_literal() {
+    /// Emit a warning if the given expressions should not be discarded.
+    /// e.g. because it's a literal (why was it made in the first place?)
+    /// e.g. because it's of the `Result` type (errors should be handled)
+    fn expression_discarded(&mut self, discarded: &TypedExpr) {
+        if discarded.is_literal() {
             self.environment.warnings.push(Warning::UnusedLiteral {
-                location: first.location(),
+                location: discarded.location(),
             });
         }
-
-        let first = self.infer(first)?;
-        let then = self.infer(then)?;
-
-        if first.type_().is_result() {
+        if discarded.type_().is_result() && !discarded.is_assignment() {
             self.environment
                 .warnings
                 .push(Warning::ImplicitlyDiscardedResult {
-                    location: first.location(),
+                    location: discarded.location(),
                 });
         }
+    }
 
-        Ok(TypedExpr::Seq {
-            typ: then.type_(),
-            first: Box::new(first),
-            then: Box::new(then),
+    fn infer_seq(
+        &mut self,
+        location: SrcSpan,
+        untyped: Vec<UntypedExpr>,
+    ) -> Result<TypedExpr, Error> {
+        let count = untyped.len();
+        let mut expressions = Vec::with_capacity(count);
+        for (i, expression) in untyped.into_iter().enumerate() {
+            let expression = self.infer(expression)?;
+            // This isn't the final expression in the sequence, so call the
+            // `expression_discarded` function to see if anything is being
+            // discarded that we think shouldn't be.
+            if i < count - 1 {
+                self.expression_discarded(&expression);
+            }
+            expressions.push(expression);
+        }
+        Ok(TypedExpr::Sequence {
+            location,
+            expressions,
         })
     }
 
@@ -506,8 +535,7 @@ impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
         location: SrcSpan,
     ) -> Result<TypedExpr, Error> {
         let tuple = self.infer(tuple)?;
-
-        match tuple.type_().as_ref() {
+        match collapse_links(tuple.type_()).as_ref() {
             Type::Tuple { elems } => {
                 let typ = elems
                     .get(index as usize)
@@ -676,49 +704,21 @@ impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
         })
     }
 
-    fn infer_let(
+    fn infer_assignment(
         &mut self,
         pattern: UntypedPattern,
         value: UntypedExpr,
-        then: UntypedExpr,
         kind: AssignmentKind,
         annotation: &Option<TypeAst>,
         location: SrcSpan,
     ) -> Result<TypedExpr, Error> {
         let value = self.in_new_scope(|value_typer| value_typer.infer(value))?;
-
-        let try_value_type = self.new_unbound_var(self.environment.level);
-        let try_error_type = self.new_unbound_var(self.environment.level);
-
-        let value_typ = match kind {
-            // Ensure that the value is a result if this is a `try` binding
-            AssignmentKind::Try => {
-                let v = try_value_type.clone();
-                let e = try_error_type.clone();
-                self.unify(result(v, e), value.type_())
-                    .map_err(|e| convert_unify_error(e, value.location()))?;
-                try_value_type
-            }
-            _ => value.type_(),
-        };
-
-        let value_typ = generalise(value_typ, self.environment.level + 1);
+        let value_typ = generalise(value.type_(), self.environment.level + 1);
 
         // Ensure the pattern matches the type of the value
         let pattern =
             pattern::PatternTyper::new(self.environment, &self.hydrator, self.environment.level)
                 .unify(pattern, value_typ.clone())?;
-
-        // Check the type of the following code
-        let then = self.infer(then)?;
-        let typ = then.type_();
-
-        // Ensure that a Result with the right error type is returned for `try`
-        if kind == AssignmentKind::Try {
-            let value = self.new_unbound_var(self.environment.level);
-            self.unify(result(value, try_error_type), typ.clone())
-                .map_err(|e| convert_unify_error(e, then.try_binding_location()))?;
-        }
 
         // Check that any type annotation is accurate.
         if let Some(ann) = annotation {
@@ -731,8 +731,64 @@ impl<'a, 'b, 'c> ExprTyper<'a, 'b, 'c> {
 
         Ok(TypedExpr::Assignment {
             location,
-            typ,
+            typ: value.type_(),
             kind,
+            pattern,
+            value: Box::new(value),
+        })
+    }
+
+    fn infer_try(
+        &mut self,
+        pattern: UntypedPattern,
+        value: UntypedExpr,
+        then: UntypedExpr,
+        annotation: &Option<TypeAst>,
+        location: SrcSpan,
+    ) -> Result<TypedExpr, Error> {
+        let value = self.in_new_scope(|value_typer| value_typer.infer(value))?;
+
+        let value_type = self.new_unbound_var(self.environment.level);
+        let try_error_type = self.new_unbound_var(self.environment.level);
+
+        // Ensure that the value is a result
+        {
+            let v = value_type.clone();
+            let e = try_error_type.clone();
+            self.unify(result(v, e), value.type_())
+                .map_err(|e| convert_unify_error(e, value.location()))?;
+        };
+
+        let value_type = generalise(value_type, self.environment.level + 1);
+
+        // Ensure the pattern matches the type of the value
+        let pattern =
+            pattern::PatternTyper::new(self.environment, &self.hydrator, self.environment.level)
+                .unify(pattern, value_type.clone())?;
+
+        // Check the type of the following code
+        let then = self.infer(then)?;
+        let typ = then.type_();
+
+        // Ensure that a Result with the right error type is returned for `try`
+        {
+            let t = self.new_unbound_var(self.environment.level);
+            self.unify(result(t, try_error_type), typ.clone())
+                .map_err(|e| convert_unify_error(e, then.location()))?;
+        }
+
+        // Check that any type annotation is accurate.
+        if let Some(ann) = annotation {
+            let ann_typ = self
+                .type_from_ast(ann)
+                .map(|t| self.instantiate(t, self.environment.level, &mut hashmap![]))?;
+            self.unify(ann_typ, value_type)
+                .map_err(|e| convert_unify_error(e, value.location()))?;
+        }
+
+        Ok(TypedExpr::Try {
+            location,
+            typ,
             pattern,
             value: Box::new(value),
             then: Box::new(then),