+++ CHANGELOG.md
@@ -6,6 +6,18 @@
 - Gleam can now run eunit without an external build tool.
 - Gleam can now run an Erlang shell without an external build tool.
 - Projects without rebar3 can be generated using the `gleam-lib` template.
+
+## v0.15.1 -2021-05-07
+
+- Fixed a bug where blocks that contained try expressions could be formatted
+  incorrectly.
+
+## v0.15.0 - 2021-05-06
+
+[Release Blog Post](https://gleam.run/news/gleam-v0.15-released/)
+
+## v0.15.0-rc1 - 2021-05-05
+
 - Syntax highlighting of Gleam code in generated HTML documentation has been
   improved.
 - Fixed a bug where markdown tables in rendered HTML documentation would have
@@ -19,6 +31,20 @@
   importing the `gleam` module.
 - Empty lists can now be used in constants.
 - Compiler performance has been improved when working with lists.
+- Compiler performance has been improved when working with sequences of
+  expressions.
+- Assignments using `let` and `assert` are now expressions and no longer require
+  a following expression in their containing block. They are now themselves
+  expessions.
+- Fixed a bug where tuple indexing could incorrectly claim a tuple is not of
+  type tuple in some circumstances.
+- Glean `new` command now checks if target folder exists, if so it returns
+  an error.
+- A compile time error is now raised if a module is defined with the name `gleam`.
+- A compile time error is now raised if a module is defined with the a keyword
+  in the name.
+- New projects are generated using `gleam_stdlib` v0.15.0.
+- New projects are generated at v0.1.0.
 
 ## v0.14.4 - 2021-03-27